add redis

+++ types/types.go
@@ -47,7 +47,7 @@ type AlertStatus struct {
 
 	// For internal tracking, not exposed in the API.
 	pendingSilences []string
-	silencesVersion int
+	silencesVersion int64
 }
 
 // Marker helps to mark alerts as silenced and/or inhibited.
@@ -59,7 +59,7 @@ type Marker interface {
 	// complete set of relevant silences. If no active silence IDs are provided and
 	// InhibitedBy is already empty, it sets the provided alert to AlertStateActive.
 	// Otherwise, it sets the provided alert to AlertStateSuppressed.
-	SetActiveOrSilenced(alert model.Fingerprint, version int, activeSilenceIDs, pendingSilenceIDs []string)
+	SetActiveOrSilenced(alert model.Fingerprint, version int64, activeSilenceIDs, pendingSilenceIDs []string)
 	// SetInhibited replaces the previous InhibitedBy by the provided IDs of
 	// alerts. In contrast to SetActiveOrSilenced, the set of provided IDs is not
 	// expected to represent the complete set of inhibiting alerts. (In
@@ -86,7 +86,7 @@ type Marker interface {
 	// result is based on.
 	Unprocessed(model.Fingerprint) bool
 	Active(model.Fingerprint) bool
-	Silenced(model.Fingerprint) (activeIDs, pendingIDs []string, version int, silenced bool)
+	Silenced(model.Fingerprint) (activeIDs, pendingIDs []string, version int64, silenced bool)
 	Inhibited(model.Fingerprint) ([]string, bool)
 }
 
@@ -151,7 +151,7 @@ func (m *memMarker) Count(states ...AlertState) int {
 }
 
 // SetActiveOrSilenced implements Marker.
-func (m *memMarker) SetActiveOrSilenced(alert model.Fingerprint, version int, activeIDs, pendingIDs []string) {
+func (m *memMarker) SetActiveOrSilenced(alert model.Fingerprint, version int64, activeIDs, pendingIDs []string) {
 	m.mtx.Lock()
 	defer m.mtx.Unlock()
 
@@ -241,7 +241,7 @@ func (m *memMarker) Inhibited(alert model.Fingerprint) ([]string, bool) {
 // Silenced returns whether the alert for the given Fingerprint is in the
 // Silenced state, any associated silence IDs, and the silences state version
 // the result is based on.
-func (m *memMarker) Silenced(alert model.Fingerprint) (activeIDs, pendingIDs []string, version int, silenced bool) {
+func (m *memMarker) Silenced(alert model.Fingerprint) (activeIDs, pendingIDs []string, version int64, silenced bool) {
 	s := m.Status(alert)
 	return s.SilencedBy, s.pendingSilences, s.silencesVersion,
 		s.State == AlertStateSuppressed && len(s.SilencedBy) > 0
